+++ tests/lantern_tests/outputs/test_site_resources.py
@@ -47,3 +47,8 @@ class TestSiteResourcesOutput:
         output = SiteResourcesOutput(logger=fx_logger, meta=fx_export_meta)
         checks = output.checks
         assert len(checks) == 3
+
+    def test_invalidation_keys(self, fx_logger: logging.Logger, fx_export_meta: ExportMeta):
+        """Can generate invalidation paths for content."""
+        output = SiteResourcesOutput(logger=fx_logger, meta=fx_export_meta)
+        assert output.invalidation_keys == ["/static/*"]